Decompose types
